xignzi006 发布系统代码 [后台服务体系，iBiz软件生产管理]

+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/ibiz/service/impl/IbzCaseServiceImpl.java
@@ -134,7 +134,9 @@ public class IbzCaseServiceImpl extends ServiceImpl<IbzCaseMapper, IbzCase> impl
     @Override
     @Transactional
     public boolean remove(Long key) {
-        ibzlibcasestepsService.removeByIbizcase(key) ;
+        if(!ibzcaseRuntime.isRtmodel()){
+            ibzlibcasestepsService.removeByIbizcase(key) ;
+        }
         boolean result = removeById(key);
         return result ;
     }
@@ -153,7 +155,9 @@ public class IbzCaseServiceImpl extends ServiceImpl<IbzCaseMapper, IbzCase> impl
             throw new BadRequestAlertException("数据不存在", this.getClass().getSimpleName(), String.valueOf(key));
         }
         else {
-            et.setIbzlibcasesteps(ibzlibcasestepsService.selectByIbizcase(key));
+        if(!ibzcaseRuntime.isRtmodel()){
+                et.setIbzlibcasesteps(ibzlibcasestepsService.selectByIbizcase(key));
+        }
         }
         return et;
     }
